chore: Some fixes and cleanup
