+++ portal/next-js/app/page.tsx
@@ -2,7 +2,7 @@ import Alert from '@components/Alert/alert';
 import { AlertClasses } from 'types/enum';
 
 const Index = () => {
-  return <Alert state={AlertClasses.PRIMARY} message={'Alerts are working'} />;
+  return <></>;
 };
 
 export default Index;
