+++ portal/next-js/app/view/admin/console/console.tsx
@@ -1,0 +1,7 @@
+interface AdminConsoleProps {}
+
+const AdminConsole = () => {
+  return <>Admin console!</>;
+};
+
+export default AdminConsole;
