+++ portal/next-js/app/view/admin/reports-form/reports-form.tsx
@@ -1,0 +1,7 @@
+interface ReportsFormProps {}
+
+const ReportsForm = () => {
+  return <>Reports form!</>;
+};
+
+export default ReportsForm;
